feat(CP): dynamic timeout calculation, osdp_cp_refresh_ex, is_pd_online
* implement dynamic response timeout based on baud rate and packet size
* add `osdp_cp_refresh_ex()` for improved PD with online/offline awareness
* expose `is_pd_online()` helper to check PD status per index

feat(pd): support alive/polling commands and improve command handling

* handle CMD_POLL as OSDP_CMD_POLLING via callback
* treat OSDP_ERR_PKT_SKIP as OSDP_CMD_ALIVE and invoke callback
* update timeout logic to use MAX(OSDP_RESP_TOUT_MS, (len * OSDP_RESP_TOUT_K * 1000) / baud_rate)
* introduce OSDP_CMD_ALIVE and OSDP_CMD_POLLING command types

+++ src/osdp_cp.c
@@ -889,7 +889,7 @@ static int cp_phy_state_update(struct osdp_pd *pd)
 			cp_phy_state_wait(pd, OSDP_CMD_RETRY_WAIT_MS);
 			return OSDP_CP_ERR_CAN_YIELD;
 		}
-        if (osdp_millis_since(pd->phy_tstamp) > MAX(OSDP_RESP_TOUT_MS, (pd->packet_buf_len * OSDP_RESP_TOUT_K * 1000) / pd->baud_rate)) {
+        if (osdp_millis_since(pd->phy_tstamp) > (int64_t)MAX(OSDP_RESP_TOUT_MS, (pd->packet_buf_len * OSDP_RESP_TOUT_K * 1000) / pd->baud_rate)) {
 			if (pd->phy_retry_count < OSDP_CMD_MAX_RETRIES) {
 				pd->phy_retry_count += 1;
 				LOG_WRN("No response in %dms; probing (%d)",
